Supression de short

+++ src/soprasteria/code24h/marabunta/communication/fourmiliere/ActionsFourmiliere.java
@@ -28,7 +28,7 @@ public class ActionsFourmiliere {
 	 * @param m0 Mémoire initiale de la fourmi. (0-255)
 	 * @param m1 Mémoire initiale de la fourmi. (0-255)
 	 */
-	public void AntOut(Integer type, Integer food, Short m0, Short m1) {
+	public void AntOut(Integer type, Integer food, Integer m0, Integer m1) {
 		this.tw.add("ANT_OUT " + type + " " + food + " " + m0 + " " + m1);
 		this.sendActions();
 	}
